% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsaiRefClass.R
\docType{class}
\name{bonsai-class}
\alias{bonsai}
\alias{bonsai-class}
\title{bonsai reference class.}
\description{
bonsai reference class.
}
\section{Methods}{

\describe{
\item{\code{automaticBurnin(verbose = TRUE)}}{Performs automatic burnin.}

\item{\code{makeRMarkdown()}}{Generates an HTML from the bonsai object.}

\item{\code{makeTeX()}}{Generates a pdf from the bonsai object.}
}}

