% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{bonsai-package}
\alias{bonsai}
\alias{bonsai-package}
\title{bonsai: Model Wrappers for Tree-Based Models}
\description{
Bindings for additional tree-based model engines for use with the 'parsnip' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bonsai.tidymodels.org}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Daniel Falbel \email{dfalbel@curso-r.com}
  \item Athos Damiani \email{adamiani@curso-r.com}
  \item Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})
  \item Simon Couch \email{simonpatrickcouch@gmail.com} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
