% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{tree_fit}
\alias{tree_fit}
\title{Internal function wrappers}
\usage{
tree_fit(
  formula,
  data,
  cost_complexity = 0.01,
  tree_depth = 30L,
  min_n = 2L,
  ...
)
}
\description{
These functions are not intended for package users.
}
\keyword{internal}
