% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{predict_catboost_regression_numeric}
\alias{predict_catboost_regression_numeric}
\title{Internal functions}
\usage{
predict_catboost_regression_numeric(object, new_data, ...)
}
\description{
Not intended for direct use.
}
\keyword{internal}
