% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CladeParameterRefClass.R
\docType{class}
\name{CladeParameter-class}
\alias{CladeParameter}
\alias{CladeParameter-class}
\title{Clade parameter reference class.}
\description{
Clade parameter reference class.
}
\section{Fields}{

\describe{
\item{\code{name}}{The parameter's name.}

\item{\code{has_posterior}}{Logical indicating whether there are samples from the posterior for the parameter.}

\item{\code{posterior_num_runs}}{The number of runs under the posterior (if any).}

\item{\code{posterior_num_generations}}{The number of samples taken from the posterior (after burnin) for each run.}

\item{\code{posterior_num_burnin}}{The number of samples taken from the posterior (before burnin) for each run.}

\item{\code{posterior_samples}}{List of samples from the posterior (after burnin). Each list element is a set of samples from a particular MCMC run.}

\item{\code{posterior_burnin_samples}}{List of samples from the posterior (before burnin). Each list element is a set of samples from a particular MCMC run.}

\item{\code{posterior_mean}}{The estimated mean of the parameter's marginal posterior distribution for each run.}

\item{\code{posterior_credible_interval}}{The estimated credible interval of the parameter's marginal posterior distribution for each run.}

\item{\code{posterior_ess}}{The effective number of samples taken from the parameter's marginal posterior distribution for each run.}

\item{\code{posterior_geweke}}{The test-statistic for Geweke's diagnostic for the parameter's posterior samples for each run.}

\item{\code{has_prior}}{Logical indicating whether there are samples from the prior for the parameter.}

\item{\code{prior_num_runs}}{The number of runs under the prior (if any).}

\item{\code{prior_num_generations}}{The number of samples taken from the prior (after burnin) for each run.}

\item{\code{prior_num_burnin}}{The number of samples taken from the prior (before burnin) for each run.}

\item{\code{prior_samples}}{List of samples from the prior (after burnin). Each list element is a set of samples from a particular MCMC run.}

\item{\code{prior_burnin_samples}}{List of samples from the prior (before burnin). Each list element is a set of samples from a particular MCMC run.}

\item{\code{prior_mean}}{The estimated mean of the parameter's marginal prior distribution for each run.}

\item{\code{prior_credible_interval}}{The estimated credible interval of the parameter's marginal prior distribution for each run.}

\item{\code{prior_ess}}{The effective number of samples taken from the parameter's marginal prior distribution for each run.}

\item{\code{prior_geweke}}{The test-statistic for Geweke's diagnostic for the parameter's prior samples for each run.}
}}
\section{Methods}{

\describe{
\item{\code{boxplots(col, posterior = TRUE, axis.lty = 1, yaxs = "r", ...)}}{Boxplots of the marginal posterior distributions for the parameter.}

\item{\code{cleanup()}}{Ensures that all the parameter summaries and diagnostics are up-to-date.}

\item{\code{densities(col, type = "p", posterior = TRUE, num_bins = 50, ...)}}{Marginal posterior densities for the parameter.}

\item{\code{initialize(name, species = NULL, posterior_samples = NULL,
  prior_samples = NULL, burnin = 0.25, ...)}}{Create an object of class CladeParameter.}

\item{\code{traces(col, posterior = TRUE, plot_burnin = TRUE, ...)}}{Plot a time-series of the parameter values.}
}}

