% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{predict_lightgbm_classification_prob}
\alias{predict_lightgbm_classification_prob}
\alias{predict_lightgbm_classification_class}
\alias{predict_lightgbm_classification_raw}
\alias{predict_lightgbm_regression_numeric}
\alias{multi_predict._lgb.Booster}
\title{Internal functions}
\usage{
predict_lightgbm_classification_prob(object, new_data, ...)

predict_lightgbm_classification_class(object, new_data, ...)

predict_lightgbm_classification_raw(object, new_data, ...)

predict_lightgbm_regression_numeric(object, new_data, ...)

\method{multi_predict}{`_lgb.Booster`}(object, new_data, type = NULL, trees = NULL, ...)
}
\description{
Not intended for direct use.
}
\keyword{internal}
