% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_package.R
\docType{package}
\name{bonsai-package}
\alias{bonsai-package}
\alias{bonsai}
\title{bonsai: Model Wrappers for Tree-Based Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bindings for additional tree-based model engines for use with the 'parsnip' package. Models include gradient boosted decision trees with 'LightGBM' (Ke et al, 2017.), conditional inference trees and conditional random forests with 'partykit' (Hothorn and Zeileis, 2015. and Hothorn et al, 2006. \doi{10.1198/106186006X133933}), and accelerated oblique random forests with 'aorsf' (Jaeger et al, 2022 \doi{10.5281/zenodo.7116854}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bonsai.tidymodels.org/}
  \item \url{https://github.com/tidymodels/bonsai}
  \item Report bugs at \url{https://github.com/tidymodels/bonsai/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Daniel Falbel \email{dfalbel@curso-r.com}
  \item Athos Damiani \email{adamiani@curso-r.com}
  \item Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
